Add Checksum protocol and simplify CRC verification
