+++ Sources/CRC32.swift
@@ -17,7 +17,6 @@ extension CRC32 {
     public static let posix = Self(polynomial: 0x04C11DB7, xorOut: 0xFFFFFFFF)
     public static let sata = Self(polynomial: 0x04C11DB7, initialValue: 0x52325032)
     public static let xfer = Self(polynomial: 0x000000AF)
-
     public static let c = Self(polynomial: 0x1EDC6F41, initialValue: 0xFFFFFFFF, reflected: true, xorOut: 0xFFFFFFFF)
     public static let d = Self(polynomial: 0xA833982B, initialValue: 0xFFFFFFFF, reflected: true, xorOut: 0xFFFFFFFF)
     public static let q = Self(polynomial: 0x814141AB)
